requirements.txt
botbuilder-integration-aiohttp>=4.14.0
jira>=3.5
